+++ src/self-check.ts
@@ -1,3 +1,4 @@
+import * as os from 'os';
 import { Logger } from './core/logger/logger.service';
 
 const logger = new Logger();
@@ -5,7 +6,7 @@ const logger = new Logger();
 /**
  * The purpose of this script is to check the environment before launching the UI
  */
-async function main() {
+function main() {
   // check if node-pty is built correctly
   try {
     require('node-pty-prebuilt-multiarch');
@@ -16,9 +17,36 @@ async function main() {
     logger.error('[node-pty] This could be because the installation of this plugin did not complete successfully ' +
       'or you may have recently upgraded Node.js to a new major version and have not reinstalled or rebuilt this plugin.');
     logger.error('[node-pty] This can usually be fixed by uninstalling and ' +
-      'reinstalling this plugin (exact command may vary based on your platform and setup):');
-    logger.error('[node-pty] sudo npm uninstall -g homebridge-config-ui-x');
-    logger.error('[node-pty] sudo npm install -g --unsafe-perm homebridge-config-ui-x');
+      'reinstalling this homebridge-config-ui-x.');
+
+    if ((process.env.UIX_SERVICE_MODE === '1')) {
+      if (os.platform() === 'win32') {
+        logger.error('[node-pty] From the Node.js command prompt (run as Administrator) run this command to rebuild Node.js modules:\n');
+        logger.warn('hb-service rebuild\n');
+      } else {
+        logger.error('[node-pty] From the terminal run this command to rebuild Node.js modules:\n');
+        logger.warn('sudo hb-service rebuild\n');
+      }
+      throw new Error('Node.js global modules rebuild required. See log errors above.');
+    } else {
+      if (os.platform() === 'win32') {
+        logger.error('[node-pty] From the Node.js command prompt (run as Administrator) run these commands (exact commands may vary):\n');
+        logger.warn('npm uninstall -g homebridge-config-ui-x');
+        logger.warn('npm install -g homebridge-config-ui-x\n');
+      } else if (os.platform() === 'darwin') {
+        logger.error('[node-pty] From the terminal run these commands (exact commands may vary):\n');
+        logger.warn('npm uninstall -g homebridge-config-ui-x');
+        logger.warn('npm install -g --unsafe-perm homebridge-config-ui-x');
+        logger.warn('cd $(npm -g prefix)/lib/node_modules');
+        logger.warn('npm rebuild --unsafe-perm\n');
+      } else {
+        logger.error('[node-pty] From the terminal run these commands (exact commands may vary):\n');
+        logger.warn('sudo npm uninstall -g homebridge-config-ui-x');
+        logger.warn('sudo npm install -g --unsafe-perm homebridge-config-ui-x');
+        logger.warn('cd $(sudo npm -g prefix)/lib/node_modules');
+        logger.warn('sudo npm rebuild --unsafe-perm\n');
+      }
+    }
     process.exit(1);
   }